BREAKING: Renamed an Avaliable capability to IsAvailable
