+++ dpl/things/capabilities/is_available.py
